refactor: initiate dispute modal after completing a contract
